requirements.txt
openai==1.3.0
python-dotenv==1.0.0
markdown==3.5.1
tiktoken==0.5.1
sentence-transformers==2.2.2
tqdm==4.65.0
streamlit==1.22.0